% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{substr_dummify}
\alias{substr_dummify}
\title{Create Dummy Variables From Substrings}
\usage{
substr_dummify(column, prefix, options, dummyVarNames, otherCol = "OTHER")
}
\arguments{
\item{column}{the column of variables to dummify}

\item{prefix}{the prefix of the dummy variable columns}

\item{options}{the substrings that we would like to extract.}

\item{dummyVarNames}{the dummy variable names (should correspond with options)}
}
\value{
a dataframe containing the column split out
}
\description{
Extract substrings from a given column, creating
dummy variables in the format of colName.DUMMY_NAME,
where DUMMY_NAME is the corresponding value from
(@param dummyVarNames), for a given option.
}
\details{
Provide this function a dataframe to work with,
a column name, the substrings that you would like extract from
the column, and a list that corresponds to what those substring's
should map to in the dummy variables.

Strips away punctation _around_ substrings, and provides an other column back out.

Note: This package matches by a pattern of "longest-first",
so "anteater" is more important to match than "ant", but if "anteater"
is unavailable, then "ant" is matched.
}
\examples{

df <- data.frame(
 a = c("ant", "anteater", "bear", "no"),
 b = c("anteater", "ant", "hi", "bear")
)

df |> mutate(substr_dummify(a, "animal", b, b))

}
